+++ packages/fm-assessment-engine/src/bootstrap.tsx
@@ -1,12 +1,12 @@
 import 'regenerator-runtime/runtime';
 import * as React from 'react';
 import { render } from 'react-dom';
-import configureStore, { initialAppState } from '@jimmy-james/store';
+import configureStore, { initialState } from '@jimmy-james/store';
 import { Provider as ReduxProvider } from 'react-redux';
 import { GlobalProvider } from '@coreym/benchmark';
 import App from './App';
 
-const store = configureStore(initialAppState);
+const store = configureStore(initialState);
 
 function initializeReactApp(): void {
   const appContainer = document.getElementById('appContainer');