+++ packages/store/src/actionTypes/assessmentActionTypes.ts
@@ -1,2 +1,0 @@
-export const LOAD_PREV_ITEM = 'LOAD_PREV_ITEM';
-export const LOAD_NEXT_ITEM = 'LOAD_NEXT_ITEM';