+++ packages/store/src/actionTypes/responses/singleSelectActionTypes.ts
@@ -1,3 +1,0 @@
-export const MCSS_CLICK = 'MCSS_CLICK';
-export const MCSS_ELIMINATE = 'MCSS_ELIMINATE';
-export const MCSS_CLEAR = 'MCSS_CLEAR';