+++ packages/store/src/actionTypes/responses/zonesActionTypes.ts
@@ -1,2 +1,0 @@
-export const ZONES_SELECT = 'ZONES_SELECT';
-export const ZONES_CLEAR = 'ZONES_CLEAR';