+++ packages/store/src/interfaces/AppLoading.ts
@@ -1,3 +1,0 @@
-export interface AppLoading {
-  apiCallsInFlightCount: number;
-}