+++ packages/store/src/interfaces/Assessment.ts
@@ -1,4 +1,0 @@
-export interface Assessment {
-  activeBlockSequence: number;
-  activeItemSequence: number;
-}