+++ packages/store/src/interfaces/Auth.ts
@@ -1,7 +1,0 @@
-export interface JWT {
-  token?: string;
-}
-
-export interface AuthState {
-  token?: string | null | undefined;
-}