+++ packages/store/src/interfaces/ErrorState.ts
@@ -1,4 +1,0 @@
-export interface ErrorState {
-  message?: string;
-  httpResponseCode?: string;
-}