+++ packages/store/src/interfaces/Item.ts
@@ -1,5 +1,0 @@
-import { RootItemConfig } from './renderer/UIElement';
-
-export interface Item {
-  item: RootItemConfig;
-}