+++ packages/store/src/interfaces/ProtectedRouteProps.ts
@@ -1,5 +1,0 @@
-export interface ProtectedRouteProps {
-  path: string;
-  component: React.FC;
-  redirectPath?: string;
-}