+++ packages/store/src/interfaces/Responses.ts
@@ -1,5 +1,0 @@
-export interface Responses {
-  [inputId: string]: {
-    [x: string]: string | number | boolean | null | string[];
-  };
-}