+++ packages/store/src/interfaces/Toolbar.ts
@@ -1,5 +1,0 @@
-// these properties are from @corey/benchmark
-export interface Toolbar {
-  scratchOpen: boolean;
-  timerOpen: boolean;
-}