+++ packages/store/src/reducers/index.ts
@@ -1,30 +1,8 @@
 import { combineReducers } from 'redux';
-import appLoadingReducer from './appLoadingReducer';
-import assessmentReducer from './assessmentReducer';
-import bookletsReducer from './bookletsReducer';
-import itemReducer from './itemReducer';
-import itemHeadersReducer from './itemHeadersReducer';
-import responsesReducer from './responsesReducer';
-// import toolbarReducer from './toolbarReducer';
-import toolSettingsReducer from './toolSettingsReducer';
-import evenCounterReducer from './OddEvenCounterReducers/evenCounterReducer';
-import oddCounterReducer from './OddEvenCounterReducers/oddCounterReducer';
-import errorReducer from './APIErrorReducer';
+import { CounterReducer } from './counterReducer';
 
 const rootReducer = combineReducers({
-  appLoadingState: appLoadingReducer,
-  assessmentState: assessmentReducer,
-  itemState: itemReducer,
-  itemHeadersState: itemHeadersReducer,
-  responsesState: responsesReducer,
-  // toolbarState: toolbarReducer,
-  evenCounterState: evenCounterReducer,
-  oddCounterState: oddCounterReducer,
-  bookletState: bookletsReducer,
-  errorState: errorReducer,
-  toolSettings: toolSettingsReducer,
-
-  // rootItemConfig: itemsReducer,
+  counterState: CounterReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;