+++ packages/store/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 0.3.0 (2020-11-05)
+
+
+### Features
+
+* redux monorepo setup with assessment-engine ([a772a6c](https://github.com/jimmy-james/fm-monorepo-poc/commit/a772a6cc0164ea1fee70a3cf1584265407ccce3a))
+* redux monorepo setup with assessment-engine ([53864d2](https://github.com/jimmy-james/fm-monorepo-poc/commit/53864d2b755a4865f48e0f6308a0f4ba9bf5f8ba))
+* redux monorepo setup with assessment-engine ([ff40be4](https://github.com/jimmy-james/fm-monorepo-poc/commit/ff40be400fd53e422cf1f2da2118c064300ad529))
+* setup redux store as monorepo ([ba47b8b](https://github.com/jimmy-james/fm-monorepo-poc/commit/ba47b8b1467da154154e8646787a75554ffb5e57))
+
+
+
+
+
 # 0.2.0 (2020-11-05)